Add custom JSON logger
I decided to implement my own logger. Perhaps not the best choice when
it comes to performance, but more valuable as a learning exercise.

A nice feature is that we define a logger interface, to allow
substitution of different implementations. Overkill for now, but it does
allow us to pass a null logger in the tests.