Experiment: test against an HTTP server
Use httptest.NewServer to start a server and send requests to it in
tests.

These aren't true end-to-end tests, because we're not running the
application itself. As our single HTTP handler is effectively
responsible for routing, there isn't really much advantage to testing
the outer code anyway. That's probably an argument for just testing the
handler directly.