Remove gin
gin isn't really necessary for this project. The object is to learn Go,
so we should lean on the standard library. Now we're feeling more
confident, gin can be removed.

We explicitly initialise a new muxer to avoid multiple handler
registration errors in tests. This might indicate that we need to tweak
our testing strategy.

To access the status in the logger we need to wrap the response writer
with a type that caches that status when written. Explicitly write the
status in all cases so that it can be logged.

Because all URLs are currently handled by a single handler, we could get
away with passing that handler directly to ListenAndServe, effectively
bypassing any routing. This works because ListenAndServe takes a handler
(the muxer implements this interface). Still, it seems more conventional
to pass a muxer, and there are clearly some benefits:

> ServeMux also takes care of sanitizing the URL request path and the
> Host header, stripping the port number and redirecting any request
> containing . or .. elements or repeated slashes to an equivalent,
> cleaner URL.

https://pkg.go.dev/net/http#ServeMux