Refactor application type
The purpose of the application type is to encapsulate dependencies,
currently a store and logger, so that these can easily be passed to the
application and injected in tests.

Previously, the type also held a reference to the muxer and acted as a
HTTP handler, which felt like a stretch of its responsibilities.

Refactor the application type and calling code to remove the reference
to the muxer. Avoid the application type needing to implement the HTTP
handler interface by defining a function that returns a handler that
closes over it.