Reorganise project files
There doesn't seem to be a generally accepted convention for structuring
a Go project. That said, common recommendations are to use a cmd
directory for commands, and an internal directory for internal packages.

One advantage of extracting packages is that it forces us to make the
contracts between components explicit. See, for example, the changes to
tests made necessary by restricting access to the store's underlying
map.

I'm not completely confident in all the package distinctions. Extracting
the handler and middleware packages is potentially premature, and the
latter feels like it could end up becoming a bit of a junk drawer.

https://github.com/golang-standards/project-layout/issues/117
https://go.dev/blog/organizing-go-code
https://go.dev/doc/go1.4#internalpackages
https://rakyll.org/style-packages/
https://eli.thegreenplace.net/2019/simple-go-project-layout-with-modules/