Simplify test assertions
There is a cost incurred when converting from a byte slice to a string.

We stream bytes to and from a HTTP request and response, so it makes
sense to store our values as byte slices to avoid incurring this cost.

Previously we were converting byte slices to strings in our tests for
comparison. Stop doing this, because it's simpler to just assert against
byte slices.

https://stackoverflow.com/questions/14373634/how-expensive-is-bytestring