Handle ignored errors
Handle previously ignored errors that could potentially occur when
reading the request body or marshalling a JSON log entry.

In the former case, I believe an error could occur when the network
connection fails or is terminated, or the request body is too large. [0]

For both those cases, it seems appropriate to panic with the error, and
let the recovery middleware handle it and return a 500.

The docs suggest that json.Marshal can return an error when called with
certain types, e.g. functions, or a cyclic data structure. [1] I don't
think that this is possible with our code, but logging out the error
feels like a solid approach regardless.

[0]: https://stackoverflow.com/a/52879005
[1]: https://pkg.go.dev/encoding/json#Marshal